Fix some java level upgrade issues

+++ core/src/main/java/com/alibaba/fastjson2/reader/FieldReaderDateField.java
@@ -82,7 +82,7 @@ final class FieldReaderDateField<T> extends FieldReaderObjectField<T> {
             value = getObjectReader(jsonReader)
                     .readObject(jsonReader);
         } else if (value instanceof Integer) {
-            long millis = ((Integer) value).intValue();
+            long millis = (Integer) value;
             if (dateReader.formatUnixTime) {
                 millis *= 1000;
             }
